Refactor  validator test to new file
