+++ app/tests/test_utils.py
@@ -8,8 +8,6 @@ import pandas as pd
 from unittest.mock import patch
 import numpy as np
 from unittest import TestCase
-from django.core.exceptions import ValidationError
-from blog.validators import snippet_validator, max_length
 
 
 class TestUtils(TestCase, MiddlewareMixin):
@@ -82,15 +80,3 @@ class TestUtils(TestCase, MiddlewareMixin):
             question="What is the capital of the United States of America?", df=df
         )
         self.assertEqual(answer, "The mocked answer")
-
-    def test_snippet_validation(self):
-        valid_value_with_links = " ".join(
-            [f'<a href="http://example{i}.com">Link{i}</a>' for i in range(100)]
-        )
-        invalid_value = f"""A {'<a href="http://example.com">Link</a> ' * 10}{'B' * (max_length - 10)}"""
-
-        with self.assertRaises(ValidationError):
-            snippet_validator(invalid_value)
-
-        snippet_validator(valid_value_with_links)
-        self.assertTrue(True)
